+++ src/Method/WebAuth.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\User;
+namespace Yiisoft\User\Method;
 
 use Psr\Http\Message\ResponseFactoryInterface;
 use Psr\Http\Message\ResponseInterface;
@@ -10,13 +10,12 @@ use Psr\Http\Message\ServerRequestInterface;
 use Yiisoft\Auth\AuthenticationMethodInterface;
 use Yiisoft\Auth\IdentityInterface;
 use Yiisoft\Http\Status;
+use Yiisoft\User\CurrentUser;
 
 /**
- * Implementation of the authentication interface for the user.
- *
- * @deprecated Use {@see \Yiisoft\User\Method\WebAuth}. Will remove in the next major version.
+ * Implementation of the `AuthenticationMethodInterface` for authenticating users in the web applications.
  */
-final class UserAuth implements AuthenticationMethodInterface
+final class WebAuth implements AuthenticationMethodInterface
 {
     private string $authUrl = '/login';